Promiscuous option, untested, -> don't merge <-.
This was prompted by issue #1 from @gabrik.

Code was borrowed from PR #12 but needed some extra work so I'm reworking it.

Basically there are uses for using the file descriptor in promiscuous mode.

BPF is easy, the kernel keeps track of the file descriptors and as soon as all
of them are closed, the interface state promiscuous mode is restored.

Linux is a bit worse, the kernel can't track the state as it uses a generic
SOCK_DGRAM socket to set IFFLAGS directly, so we have to be careful to unset the
flag once the user does a Rawlink.close_link. This will also imply that the user
must carefully add "atexit" callbacks to guarantee safe exit, the library should
not do this, but we should document it carefully.

Libraries that silently add "atexit" hooks are baaaaaaaad.

Lwt parts still missing, code completely untested.