Implement promiscuous option for open_link.
This was prompted by issue #1 from @gabrik.

This allows one to put the device in promiscuous mode when opening the link.

On linux (af_packet) we can use PACKET_MR_PROMISC which is nicer than
SIOCSIFFLAGS, that is because the kernel keeps state of the process that joined
the membership and clears it when the process dies, that's also what libpcap(3)
does. I struggled a bit in the beginning since I was doing SIOCIFFLAGS and had
to keep the state in userland which is almost impossible to do it right if we
get a SIGKILL.

BPF always did this, so it's also straightforward.